+++ src/main/java/lotto/domain/dto/request/NumberOfTicketsRequest.java
@@ -1,4 +1,4 @@
-package lotto.domain.dto;
+package lotto.domain.dto.request;
 
 public class NumberOfTicketsRequest {
     private final int purchaseAmount;
