+++ src/main/java/lotto/domain/dto/response/LottoNumbersResponse.java
@@ -1,4 +1,4 @@
-package lotto.domain.dto;
+package lotto.domain.dto.response;
 
 import lotto.domain.LottoNumber;
