+++ src/test/java/lotto/domain/LottoRankTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.enums.LottoRank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
