+++ src/test/java/lotto/domain/WinningStatisticsTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.enums.LottoRank;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
